+++ src/MaterialDatatableBody.js
@@ -90,7 +90,7 @@ class MaterialDatatableBody extends React.Component {
   render() {
     const { classes, columns, options } = this.props;
     const tableRows = this.buildRows();
-    
+
     return (
       <TableBody>
         {tableRows ? (